Move Zuora error modules to errors directory
Renamed zuoraError.ts and zuoraErrorHandler.ts to the new errors subdirectory for better organization. Updated import paths in zuoraClient.ts, zuoraErrorHandler.ts, and zuoraClient.test.ts to reflect the new locations.
